fix: sequence equal check failed

+++ leetcode/401_binary_watch.go
@@ -21,7 +21,7 @@ func readBinaryWatch(turnedOn int) []string {
 	ret := make([]string, 0)
 	for hour := 0; hour < 12; hour++ {
 		for minute := 0; minute < 60; minute++ {
-			if utils.BitCountUint32(uint32(hour))+utils.BitCountUint32(uint32(minute)) == uint32(turnedOn) {
+			if utils.BitCountUint32(uint32(hour))+utils.BitCountUint32(uint32(minute)) == turnedOn {
 				if minute < 10 {
 					ret = append(ret, fmt.Sprintf("%d:0%d", hour, minute))
 				} else {
